backend/pyproject.toml
[project]
name = "careagents-backend"
version = "1.0.0"
description = "Healthcare Management System Backend with PostgreSQL"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.0",
    "python-multipart==0.0.12",
    "asyncpg==0.29.0",
    "pydantic[email]==2.9.2",
    "pydantic-settings==2.6.0",
    "python-dotenv==1.0.1",
    "dotenv>=0.9.9",
    # AI Agent dependencies
    "anthropic>=0.39.0",
    "pyyaml>=6.0.1",
    "websockets>=12.0",
    "redis>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "httpx==0.27.2",
    "black==24.10.0",
    "ruff==0.7.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"
